Add files via upload
Added new files that are needed for the script. Now the script is set to run in a docker container with scheduled run time using cron expressiosn and the csv file in google drive has to get updates according to it. 

Also to check if the script works properly use in ubuntu in the folder '/mnt/c/Users/darko/Ubuntu_Projects/ETL_test.py'  (or where you have downloaded the script) run the command "docker run my-spotify-script" and the script will run and work. 

If there are some changes made to the code in the ('/mnt/c/Users/darko/Ubuntu_Projects/ETL_test.py') file, then first run the command "docker build -t my-spotify-script . " and the image will get a rebuild. After that use the command above and the script will run and work. 

To see if the script is running properly and if the cron is working use docker desktop, open the container that it is running into and then go to Logs. There you have to see the message 'Cron is working'.